+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Development version
 
+* Make the duplicate search more reliable when running it multiple times without closing the application in between.
 
 # Version 0.5.1 - 2020-06-19